NetskopeTransaction/pyproject.toml
[tool.poetry]
name = "Netskope Transaction events automation module"
description = ""
version = "0.0"
authors = []

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
sekoia-automation-sdk = { version="^1.8.2", extras=["all"] }
google-cloud-pubsublite = "^1.8.3"
google-api-python-client = "^2.95.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
google-api-core = "^2.11.1"
google-auth = "^2.22.0"
googleapis-common-protos = "^1.59.1"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-cov = "*"

[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 119

[tool.coverage.run]
omit = [
    "tests/*",
    "dev.py",
    "main.py",
]
